requirements.txt
alembic==1.0.7
asn1crypto==0.24.0
bcrypt==3.1.4
blinker==1.4
bulma==1.0.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.2
cycler==0.10.0
decorator==4.3.2
dominate==2.3.5
elasticsearch==6.3.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.1
flask-ponywhoosh==1.0.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
kiwisolver==1.0.1
Mako==1.0.7
MarkupSafe==1.0
matplotlib==3.0.2
nbformat==4.4.0
ndg-httpsclient==0.5.1
numpy==1.16.1
pandas==0.24.1
peewee==3.8.2
Pillow==5.3.0
plotly==3.6.0
pony==0.7.9
ponywhoosh==1.7.8
pusher==2.1.2
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
pyOpenSSL==18.0.0
pyparsing==2.3.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2018.9
requests==2.21.0
retrying==1.3.3
six==1.11.0
SQLAlchemy==1.2.12
traitlets==4.3.2
urllib3==1.24.1
virtualenv==16.0.0
visitor==0.1.3
Werkzeug==0.14.1
Whoosh==2.7.4
WhooshAlchemy==0.3.1
WTForms==2.2.1
